+++ src/context/AuthProvider.js
@@ -1,19 +1,12 @@
 import { useModal } from '@/hooks/useModal';
-import {
-  createLogin,
-  createLogout,
-  createUser,
-  getUserMe,
-} from '@/service/api/auth';
+import { createLogin, createUser, getUserMe } from '@/service/api/auth';
 import { useMutation, useQueryClient, useQuery } from '@tanstack/react-query';
-import { useRouter } from 'next/router';
-import { createContext, useContext, useEffect, useState } from 'react';
+import { createContext, useEffect, useState } from 'react';
 
-const AuthContext = createContext({
+export const AuthContext = createContext({
   user: null,
   isLoading: false,
   login: () => {},
-  logout: () => {},
 });
 
 export function AuthProvider({ children }) {
@@ -65,16 +58,6 @@ export function AuthProvider({ children }) {
     },
   });
 
-  const logoutMutation = useMutation({
-    mutationFn: createLogout,
-    onSuccess: (data) => {
-      onModalOpen({ msg: data.message });
-      localStorage.removeItem('accessToken');
-      queryClient.setQueriesData(['user'], null);
-      router.push('/');
-    },
-  });
-
   return (
     <AuthContext.Provider
       value={{
@@ -82,30 +65,10 @@ export function AuthProvider({ children }) {
         isLoading,
         login: loginMutation,
         signUp: signUpMutation,
-        logout: logoutMutation,
       }}
     >
       {children}
       <AuthModal />
     </AuthContext.Provider>
   );
-}
-
-export function useAuth(required) {
-  const context = useContext(AuthContext);
-  const router = useRouter();
-
-  if (!context) {
-    throw new Error('Error: not used within AuthProvider');
-  }
-
-  useEffect(() => {
-    if (required && !context.user && !context.isLoading) {
-      context.onModalOpen({
-        msg: '로그인 된 유저만 접근할수 있습니다.',
-        action: () => router.push('/auth/login'),
-      });
-    }
-  }, [context.user, context.isLoading, router, required]);
-  return context;
 }
